Adds plan detection
After login, the plan list is automatically retrieved from the API to
auto-activate a plan. This is handled by the PlansActionsHandler
service, which is called after PlansActions finishes a dispatch.

After list retrieval, the service tries to choose a Plan:

Is there an active plan?
    No => Is there one in localStorage?
        No => Activate the first existing plan
        Yes => Is that one an existing plan?
            Yes => Activate it
            No => Show warning and activate first existing plan
    Yes => Is it an existing plan?
        No => Show warning and activate first existing plan the list

After a plan has been chosen, the service stores the plan in
localStorage.

Change-Id: I80affcd2ef902e2050d80cc682c27661122e7b3d